# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ChemAI Model Extractor API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /variables/{id}:
    patch:
      operationId: patchVariable
      tags: [editing]
      summary: Update one or more fields of a variable (human verification)
      description: |
        Accepts a partial update. On first edit the original extracted value is
        snapshotted into originalValue for audit purposes. rawExtractionJson is
        never mutated. Returns the full updated variable row.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchVariableInput"
      responses:
        "200":
          description: Updated variable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Variable"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Variable not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /variables/{id}/reset:
    post:
      operationId: resetVariable
      tags: [editing]
      summary: Reset a variable to its original extracted value
      description: |
        Restores all fields from the originalValue snapshot, clears editedByUser,
        and nulls originalValue. No-op if the variable has not been edited.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reset variable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Variable"
        "404":
          description: Variable not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /parameters/{id}:
    patch:
      operationId: patchParameter
      tags: [editing]
      summary: Update one or more fields of a parameter (human verification)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchParameterInput"
      responses:
        "200":
          description: Updated parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Parameter"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Parameter not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /parameters/{id}/reset:
    post:
      operationId: resetParameter
      tags: [editing]
      summary: Reset a parameter to its original extracted value
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reset parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Parameter"
        "404":
          description: Parameter not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /equations/{id}:
    patch:
      operationId: patchEquation
      tags: [editing]
      summary: Update one or more fields of an equation (human verification)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchEquationInput"
      responses:
        "200":
          description: Updated equation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equation"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Equation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /equations/{id}/reset:
    post:
      operationId: resetEquation
      tags: [editing]
      summary: Reset an equation to its original extracted value
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reset equation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equation"
        "404":
          description: Equation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /assumptions/{id}:
    patch:
      operationId: patchAssumption
      tags: [editing]
      summary: Update one or more fields of an assumption or limitation (human verification)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchAssumptionInput"
      responses:
        "200":
          description: Updated assumption/limitation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assumption"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Assumption not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /assumptions/{id}/reset:
    post:
      operationId: resetAssumption
      tags: [editing]
      summary: Reset an assumption/limitation to its original extracted value
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reset assumption/limitation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assumption"
        "404":
          description: Assumption not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Equation:
      type: object
      properties:
        id:
          type: integer
        ordinal:
          type: integer
        label:
          type: string
          description: Short label, e.g. "(1)" or "Eq. 3".
        latex:
          type: string
          description: LaTeX representation of the equation.
        plaintext:
          type: string
          description: Plain-text (ASCII) representation.
        meaning:
          type: string
          description: Human-readable explanation of what the equation represents.
        variablesInvolved:
          type: array
          items:
            type: string
          description: Symbols appearing in this equation.
        confidence:
          type: string
          enum: [high, medium, low]
        description:
          type: string
          description: Auto-generated combined description (label + meaning + plaintext). Kept for backward compatibility.
        sourceQuote:
          type: string
        editedByUser:
          type: boolean
          description: True if any field has been manually edited after extraction.
        originalValue:
          type: ["object", "null"]
          additionalProperties: true
          description: Snapshot of the row before the first human edit. Null if unedited.
      required: [id, ordinal, label, latex, plaintext, meaning, variablesInvolved, confidence, description, sourceQuote, editedByUser]
    Variable:
      type: object
      properties:
        id:
          type: integer
        ordinal:
          type: integer
        symbol:
          type: string
        name:
          type: string
        meaning:
          type: string
          description: Explanation of what the variable represents physically.
        unit:
          type: string
        role:
          type: string
          enum: [state, input, output]
        confidence:
          type: string
          enum: [high, medium, low]
        sourceQuote:
          type: string
        editedByUser:
          type: boolean
        originalValue:
          type: ["object", "null"]
          additionalProperties: true
      required: [id, ordinal, symbol, name, meaning, unit, role, confidence, sourceQuote, editedByUser]
    Parameter:
      type: object
      properties:
        id:
          type: integer
        ordinal:
          type: integer
        symbol:
          type: string
        name:
          type: string
          description: Full descriptive name of the parameter.
        value:
          type: number
        unit:
          type: string
        confidence:
          type: string
          enum: [high, medium, low]
        sourceQuote:
          type: string
        editedByUser:
          type: boolean
        originalValue:
          type: ["object", "null"]
          additionalProperties: true
      required: [id, ordinal, symbol, name, value, unit, confidence, sourceQuote, editedByUser]
    Assumption:
      type: object
      properties:
        id:
          type: integer
        ordinal:
          type: integer
        kind:
          type: string
          enum: [assumption, limitation]
        text:
          type: string
        sourceQuote:
          type: string
        confidence:
          type: string
          enum: [high, medium, low]
        editedByUser:
          type: boolean
        originalValue:
          type: ["object", "null"]
          additionalProperties: true
      required: [id, ordinal, kind, text, sourceQuote, confidence, editedByUser]
    PatchVariableInput:
      type: object
      description: Partial update for a variable. Only provided fields are changed.
      properties:
        symbol:
          type: string
          minLength: 1
        name:
          type: string
        meaning:
          type: string
        unit:
          type: string
        role:
          type: string
          enum: [state, input, output]
        confidence:
          type: string
          enum: [high, medium, low]
        sourceQuote:
          type: string
    PatchParameterInput:
      type: object
      description: Partial update for a parameter. Only provided fields are changed.
      properties:
        symbol:
          type: string
          minLength: 1
        name:
          type: string
        value:
          type: number
        unit:
          type: string
        confidence:
          type: string
          enum: [high, medium, low]
        sourceQuote:
          type: string
    PatchEquationInput:
      type: object
      description: Partial update for an equation. Only provided fields are changed.
      properties:
        label:
          type: string
        latex:
          type: string
          minLength: 1
        plaintext:
          type: string
        meaning:
          type: string
        variablesInvolved:
          type: array
          items:
            type: string
        confidence:
          type: string
          enum: [high, medium, low]
        sourceQuote:
          type: string
    PatchAssumptionInput:
      type: object
      description: Partial update for an assumption or limitation. Only provided fields are changed.
      properties:
        text:
          type: string
          minLength: 1
        kind:
          type: string
          enum: [assumption, limitation]
        sourceQuote:
          type: string
        confidence:
          type: string
          enum: [high, medium, low]
...