# lib/api-zod/src/generated/types/parameter.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ChemAI Model Extractor API
 * OpenAPI spec version: 0.1.0
 */
import type { ParameterConfidence } from "./parameterConfidence";
import type { ParameterOriginalValue } from "./parameterOriginalValue";

export interface Parameter {
  id: number;
  ordinal: number;
  symbol: string;
  /** Full descriptive name of the parameter. */
  name: string;
  value: number;
  unit: string;
  confidence: ParameterConfidence;
  sourceQuote: string;
  editedByUser: boolean;
  /** @nullable */
  originalValue?: ParameterOriginalValue;
}