# lib/api-zod/src/generated/types/patchParameterInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ChemAI Model Extractor API
 * OpenAPI spec version: 0.1.0
 */
import type { PatchParameterInputConfidence } from "./patchParameterInputConfidence";

/**
 * Partial update for a parameter. Only provided fields are changed.
 */
export interface PatchParameterInput {
  /** @minLength 1 */
  symbol?: string;
  name?: string;
  value?: number;
  unit?: string;
  confidence?: PatchParameterInputConfidence;
  sourceQuote?: string;
}