# lib/api-zod/src/generated/types/variable.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ChemAI Model Extractor API
 * OpenAPI spec version: 0.1.0
 */
import type { VariableConfidence } from "./variableConfidence";
import type { VariableOriginalValue } from "./variableOriginalValue";
import type { VariableRole } from "./variableRole";

export interface Variable {
  id: number;
  ordinal: number;
  symbol: string;
  name: string;
  /** Explanation of what the variable represents physically. */
  meaning: string;
  unit: string;
  role: VariableRole;
  confidence: VariableConfidence;
  sourceQuote: string;
  editedByUser: boolean;
  /** @nullable */
  originalValue?: VariableOriginalValue;
}